Some minor tweaks to avoid breaking changes
Noticed that not all movie data gotten from the API were populated. Some properties were either blank or empty, and when these properties were used in the application, it broke. So made some minor decision statements that checks for the needed property first, using the result to choose what is reflected on the application